# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/nango/webhook": {
      "post": {
        "tags": [
          "nango"
        ],
        "operationId": "nango_webhook",
        "responses": {
          "200": {
            "description": "Webhook processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid signature"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WebhookResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}